Force Auth::id() to be integer for strict checks
For some reason in the Dev implementation of Projectify Auth::id()
returns the user's ID as a string.  This is not the case in Production.
Rather than troubleshoot Laravel 4.2 to get Auth::id() to return
integers, we'll just cast the result to int every time we call it for
now.

+++ app/controllers/SessionsController.php
@@ -25,12 +25,12 @@ class SessionsController extends BaseController {
     {
         if (Auth::attempt(Input::only('email', 'password'), Input::has('remember')))
         {
-            if (User::find(Auth::id())->confirmed) {
+            if (User::find((int)Auth::id())->confirmed) {
                 return Redirect::to('/home');
             }
             else {
                 // User still needs to verify email address
-                Session::flash('unverifiedUserId', Auth::id());
+                Session::flash('unverifiedUserId', (int)Auth::id());
                 Auth::logout();
                 return Redirect::to('/unverified');
             }
